Extract cleanResult to util
